moved Earn aave data to different context added loaders (instead of blocking the whole view) to the earn product cards
