+++ pages/earn.tsx
@@ -1,5 +1,5 @@
 import { DeferedContextProvider } from 'components/DeferedContextProvider'
-import { aaveContext, AaveContextProvider } from 'features/earn/aave/AaveContextProvider'
+import { earnContext, EarnContextProvider } from 'features/earn/EarnContextProvider'
 import { Survey } from 'features/survey'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import React from 'react'
@@ -16,14 +16,14 @@ export const getStaticProps = async ({ locale }: { locale: string }) => ({
 
 function EarnPage() {
   return (
-    <AaveContextProvider>
-      <DeferedContextProvider context={aaveContext}>
+    <EarnContextProvider>
+      <DeferedContextProvider context={earnContext}>
         <WithConnection>
           <EarnView />
           <Survey for="earn" />
         </WithConnection>
       </DeferedContextProvider>
-    </AaveContextProvider>
+    </EarnContextProvider>
   )
 }
